Added new api for supply item

+++ PortalDB/Entities/ASSET/Delivery/TblDeliveryRecordItem.cs
@@ -68,7 +68,7 @@ namespace PortalDB.Entities.ASSET.Delivery
         }
 
         [Column("DeliveryRecordItemQuantity")]
-        public long? ItemQuantity { get; set; }
+        public int? ItemQuantity { get; set; }
 
         //[Column("DeliveryRecordItemCurrentStock")]
         //public int? CurrentStock { get; set; }
