+++ PortalDB/Models/QueryParams/Delivery/EditDeliveryRecordItemRawQueryParams.cs
@@ -14,7 +14,7 @@ namespace PortalDB.Models.QueryParams.Delivery
         [Required] public long CategoryId { get; set; }
         [Required] public string ItemDescription { get; set; } = string.Empty;
         public string ItemSpecification { get; set; } = string.Empty;
-        [Required] public long ItemQuantity { get; set; }
+        [Required] public int ItemQuantity { get; set; }
         //public int CurrentStock { get; set; }
         public int ReorderPoint { get; set; }
         [Required] public long UnitId { get; set; }
